moving files and removing client base
