+++ experimental/reporting-ui/src/main/react/reporting-ui/client/reporting/testing/BUILD.bazel
@@ -1,0 +1,30 @@
+load("@aspect_rules_js//js:defs.bzl", "js_library")
+load("@aspect_rules_webpack//webpack:defs.bzl", "webpack_devserver")
+
+js_library(
+    name = "fake_app",
+    srcs = [
+        "fake_api.ts",
+        "index.tsx",
+    ],
+    deps = [
+        "//:node_modules/react",
+        "//:node_modules/react-dom",
+        "//experimental/reporting-ui/src/main/react/reporting-ui/client",
+        "//experimental/reporting-ui/src/main/react/reporting-ui/model/reporting",
+    ],
+)
+
+webpack_devserver(
+    name = "webpack_testserver",
+    data = [
+        ":fake_app",
+        "//:tsconfig",
+        "//experimental/reporting-ui/src/main/react/reporting-ui:node_deps",
+        "//experimental/reporting-ui/src/main/react/reporting-ui:app",
+        "//experimental/reporting-ui/src/main/react/reporting-ui/model/reporting",
+        "//experimental/reporting-ui/src/main/react/reporting-ui/public",
+    ],
+    node_modules = "//:node_modules",
+    webpack_config = ":webpack.config.js",
+)
