+++ experimental/reporting-ui/src/main/react/reporting-ui/client/reporting/testing/webpack.config.js
@@ -16,7 +16,7 @@ const path = require('path');
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 
 module.exports = {
-  entry: path.resolve(__dirname, './src/index.tsx'),
+  entry: path.resolve(__dirname, './index.tsx'),
   module: {
     rules: [
       {
@@ -40,11 +40,11 @@ module.exports = {
     modules: ['node_modules'],
   },
   output: {
-    path: path.resolve(__dirname, '../../../../main/react/reporting-ui/public'),
+    path: path.resolve(__dirname, '../../../public'),
     filename: 'app.bundle.js',
   },
   plugins: [new MiniCssExtractPlugin()],
   devServer: {
-    static: path.resolve(__dirname, '../../../../main/react/reporting-ui/public'),
+    static: path.resolve(__dirname, '../../../public'),
   },
 };
