+++ experimental/reporting-ui/src/test/react/reporting-ui/BUILD.bazel
@@ -37,8 +37,8 @@ jest_test(
         ":test_files",
         ":test_node_deps",
         "//:tsconfig",
-        "//experimental/reporting-ui/src/main/react/reporting-ui:build_node_deps",
-        "//experimental/reporting-ui/src/main/react/reporting-ui:src",
+        "//experimental/reporting-ui/src/main/react/reporting-ui:node_deps",
+        "//experimental/reporting-ui/src/main/react/reporting-ui:app",
         "//experimental/reporting-ui/src/test/react/jest",
     ],
     node_modules = "//:node_modules",
