--- experimental/reporting-ui/src/test/react/reporting-ui/fake/BUILD.bazel
@@ -1,15 +1,0 @@
-load("@aspect_rules_webpack//webpack:defs.bzl", "webpack_devserver")
-
-webpack_devserver(
-    name = "webpack_testserver",
-    data = [
-        "//:tsconfig",
-        "//experimental/reporting-ui/src/main/react/reporting-ui:build_node_deps",
-        "//experimental/reporting-ui/src/main/react/reporting-ui:src",
-        "//experimental/reporting-ui/src/main/react/reporting-ui/model/reporting",
-        "//experimental/reporting-ui/src/main/react/reporting-ui/public",
-        "//experimental/reporting-ui/src/test/react/reporting-ui/fake/src",
-    ],
-    node_modules = "//:node_modules",
-    webpack_config = ":webpack.config.js",
-)
