--- experimental/reporting-ui/src/test/react/reporting-ui/fake/src/BUILD.bazel
@@ -1,21 +1,0 @@
-load("@aspect_rules_js//js:defs.bzl", "js_library")
-
-package(
-    default_visibility = [
-        "//experimental/reporting-ui/src/test/react/reporting-ui:__subpackages__",
-    ],
-)
-
-js_library(
-    name = "src",
-    srcs = [
-        "fake_api.ts",
-        "index.tsx",
-    ],
-    deps = [
-        "//:node_modules/react",
-        "//:node_modules/react-dom",
-        "//experimental/reporting-ui/src/main/react/reporting-ui/client",
-        "//experimental/reporting-ui/src/main/react/reporting-ui/model/reporting",
-    ],
-)
